Lots of small update from Vivien's own version

+++ pipe/2_skysub_scripts/1_skysub_NU.py
@@ -31,6 +31,8 @@ proquest(askquestions)
 
 starttime = datetime.now()
 
+redofromscratch = True
+
 for i,image in enumerate(images):
 	
 	print "+ " * 30
@@ -39,13 +41,17 @@ for i,image in enumerate(images):
 	imagepath = os.path.join(alidir, image["imgname"] + ".fits")
 	skyimagepath = os.path.join(alidir,  image["imgname"] + "_sky.fits")
 	if os.path.isfile(skyimagepath):
-		print "Removing existing sky image."
-		os.remove(skyimagepath)
+		if redofromscratch:
+			print "Removing existing sky image."
+			os.remove(skyimagepath)
+		else:
+			print "Already done, I skip this one..."
+			continue
 	
 	# We run sextractor  on the image in electrons :
 	saturlevel = image["gain"] * image["saturlevel"] # to convert to electrons
 	
-	if image["telescopename"] in ["EulerCAM", "SMARTSandicam"]:
+	if image["telescopename"] in ["EulerCAM", "SMARTSandicam", "GMOS"]:
 		print "Detected fine pixel telescope, switching to smooth sky"
 		cmd = "%s %s -c default_sky_template_smooth.sex -PIXEL_SCALE %.3f -SATUR_LEVEL %.3f -CHECKIMAGE_NAME %s" % (sex, imagepath, image["pixsize"], saturlevel, skyimagepath)
